reader/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
feedparser==5.2.1
idna==2.9
multidict==4.7.5
python-engineio==3.12.1
python-socketio==4.5.1
six==1.14.0
yarl==1.4.2
